+++ squadcastv1/src/squadcast/models/v3_services_overlay_overlayresponse.py
@@ -8,11 +8,7 @@ from .v3_services_overlay_dedupkeyoverlay import (
 from datetime import datetime
 from pydantic import model_serializer
 from squadcast.types import BaseModel, Nullable, UNSET_SENTINEL
-from typing import Literal
 from typing_extensions import TypedDict
-
-
-OverlayTemplateType = Literal["dedup_key",]
 
 
 class V3ServicesOverlayOverlayResponseTypedDict(TypedDict):
@@ -23,7 +19,7 @@ class V3ServicesOverlayOverlayResponseTypedDict(TypedDict):
     service_id: str
     alert_source_version: str
     alert_source_shortname: str
-    overlay_template_type: OverlayTemplateType
+    overlay_template_type: str
     overlay: V3ServicesOverlayDedupKeyOverlayTypedDict
     created_by: str
     updated_by: str
@@ -45,7 +41,7 @@ class V3ServicesOverlayOverlayResponse(BaseModel):
 
     alert_source_shortname: str
 
-    overlay_template_type: OverlayTemplateType
+    overlay_template_type: str
 
     overlay: V3ServicesOverlayDedupKeyOverlay